requirements.txt
django-user-accounts==1.3.1
pinax-webanalytics==2.0.4
pinax-eventlog==1.1.2
Django==1.10.2
requests==2.11.1
pinax-theme-bootstrap==7.11.0

djangorestframework>=3.3
pillow

# Optional
Werkzeug
django-extensions
pydes
futures
